'Fix' the test with sleep(10 minutes 10 seconds)
Update of assigned roles (`GrantedRoles`) on `User` instance that occurs
inside `LDAPAccessStorage` doesn't bubble up to `AccessStorate`
and not broadcasted as an event to `AccessChangesNotifier` event handlers.

Thus `AccessControl` that is held by `Context` never receives an event and
is never updated, holding old instance of `User` with previous set of roles.

`AccessControl` instances are cached for 600 seconds, so waiting for 600 + 10
seconds between sessions ensures that when user is logged in again, new
`AccessControl` instance is created, and proper (updated `User`, with
updated `GrantedRoles`) is created.

This issue seems to be already fixed in upstream/master,
but for now we are not going to backport it, just fix the test.